fix(puccinier): add argb support
Adjust the regex so it doesn't require a '#' in front of the hex code.
This allows it to match colors in ARGB format where the color looks like
'#ffRRGGBB' instead of '#RRGGBB(AA)?'. Also prevent the regex from
matching 3-char hex codes since none of the codes in the palette are
3-char anyway.

Fixes #19